// src/lib/api/generated/models/v1ExternalOTCContractStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1ExternalOTCContractStatus {
    EXTERNAL_OTC_CONTRACT_STATUS_UNSPECIFIED = 'EXTERNAL_OTC_CONTRACT_STATUS_UNSPECIFIED',
    EXTERNAL_OTC_CONTRACT_STATUS_ACTIVE = 'EXTERNAL_OTC_CONTRACT_STATUS_ACTIVE',
    EXTERNAL_OTC_CONTRACT_STATUS_EXERCISED = 'EXTERNAL_OTC_CONTRACT_STATUS_EXERCISED',
    EXTERNAL_OTC_CONTRACT_STATUS_EXPIRED = 'EXTERNAL_OTC_CONTRACT_STATUS_EXPIRED',
    EXTERNAL_OTC_CONTRACT_STATUS_SETTLING = 'EXTERNAL_OTC_CONTRACT_STATUS_SETTLING',
}